Add permission checking for all messages to and from mobile
