+++ src/components/application_manager/include/application_manager/commands/command_notification_from_mobile_impl.h
@@ -50,6 +50,7 @@ class CommandNotificationFromMobileImpl : public CommandImpl {
       policy::PolicyHandlerInterface& policy_handler);
   virtual ~CommandNotificationFromMobileImpl();
   virtual bool Init();
+  bool CheckPermissions() OVERRIDE;
   virtual bool CleanUp();
   virtual void Run();
   void SendNotification();
