+++ src/components/application_manager/src/commands/command_notification_from_mobile_impl.cc
@@ -57,6 +57,10 @@ bool CommandNotificationFromMobileImpl::Init() {
   return true;
 }
 
+bool CommandNotificationFromMobileImpl::CheckPermissions() {
+  return CheckAllowedParameters(Command::CommandSource::SOURCE_MOBILE);
+}
+
 bool CommandNotificationFromMobileImpl::CleanUp() {
   return true;
 }
